# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpsFlow API
  description: Self-hostable, cloud-native operations platform API Gateway specification.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local API Gateway
paths:
...
  /incidents:
    get:
      summary: List Incidents
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: severity
          schema:
            type: string
      responses:
        '200':
          description: List of incidents
    post:
      summary: Create Incident
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                severity:
                  type: string
                  enum: [SEV1, SEV2, SEV3, SEV4]
      responses:
        '201':
          description: Incident created
...
  /services:
    get:
      summary: List Services
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: owner_team
          schema:
            type: string
        - in: query
          name: criticality
          schema:
            type: string
      responses:
        '200':
          description: List of registered services
    post:
      summary: Register Service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, owner_team]
              properties:
                name:
                  type: string
                description:
                  type: string
                owner_team:
                  type: string
                criticality:
                  type: string
                  enum: [CRITICAL, HIGH, MEDIUM, LOW]
                repository_url:
                  type: string
      responses:
        '201':
          description: Service registered
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT